+++ packages/ast-spec/jest.config.js
@@ -4,5 +4,6 @@
 /** @type {import('@jest/types').Config.InitialOptions} */
 module.exports = {
   ...require('../../jest.config.base.js'),
-  testRegex: ['./tests/.+\\.test\\.ts$'],
+  collectCoverage: false,
+  setupFilesAfterEnv: ['./tests/util/setupJest.ts'],
 };