+++ packages/eslint-plugin/src/rules/no-non-null-asserted-nullish-coalescing.ts
@@ -35,7 +35,7 @@ export default util.createRule({
     docs: {
       description:
         'Disallows using a non-null assertion in the left operand of the nullish coalescing operator',
-      recommended: false,
+      recommended: 'strict',
     },
     messages: {
       noNonNullAssertedNullishCoalescing: