+++ packages/utils/src/ts-eslint/RuleTester.ts
@@ -11,6 +11,7 @@ import {
 interface ValidTestCase<TOptions extends Readonly<unknown[]>> {
   /**
    * Name for the test case.
+   * @since 8.1.0
    */
   readonly name?: string;
   /**
@@ -47,6 +48,7 @@ interface ValidTestCase<TOptions extends Readonly<unknown[]>> {
   readonly settings?: Readonly<SharedConfigurationSettings>;
   /**
    * Run this case exclusively for debugging in supported test frameworks.
+   * @since 7.29.0
    */
   readonly only?: boolean;
 }