+++ packages/utils/src/ts-eslint/SourceCode.ts
@@ -288,7 +288,6 @@ declare class SourceCodeBase extends TokenStore {
    * Determines if two nodes or tokens have at least one whitespace character
    * between them. Order does not matter. Returns false if the given nodes or
    * tokens overlap.
-   * This was added in v6.7.0.
    * @since 6.7.0
    * @param first The first node or token to check between.
    * @param second The second node or token to check between.